+++ bindingcollectionadapter-recyclerview/src/main/java/me/tatarka/bindingcollectionadapter/BindingRecyclerViewAdapters.java
@@ -1,0 +1,57 @@
+package me.tatarka.bindingcollectionadapter;
+
+import android.databinding.BindingAdapter;
+import android.databinding.BindingConversion;
+import android.support.v7.widget.RecyclerView;
+import me.tatarka.bindingcollectionadapter.factories.BindingRecyclerViewAdapterFactory;
+
+import java.util.List;
+
+/**
+ * @see {@link BindingCollectionAdapters}
+ */
+public class BindingRecyclerViewAdapters {
+    // RecyclerView
+    @SuppressWarnings("unchecked")
+    @BindingAdapter({"itemView", "items"})
+    public static <T> void setAdapter(RecyclerView recyclerView, ItemViewArg<T> arg, List<T> items) {
+        setAdapter(recyclerView, arg, items, BindingRecyclerViewAdapterFactory.DEFAULT, null);
+    }
+
+    @BindingAdapter({"itemView", "items", "adapter"})
+    public static <T> void setAdapter(RecyclerView recyclerView, ItemViewArg<T> arg, List<T> items, BindingRecyclerViewAdapterFactory factory) {
+        setAdapter(recyclerView, arg, items, factory, null);
+    }
+
+    @BindingAdapter({"itemView", "items", "itemIds"})
+    public static <T> void setAdapter(RecyclerView recyclerView, ItemViewArg<T> args, List<T> items, BindingRecyclerViewAdapter.ItemIds<T> itemIds) {
+        setAdapter(recyclerView, args, items, BindingRecyclerViewAdapterFactory.DEFAULT, itemIds);
+    }
+
+    @SuppressWarnings("unchecked")
+    @BindingAdapter({"itemView", "items", "adapter", "itemIds"})
+    public static <T> void setAdapter(RecyclerView recyclerView, ItemViewArg<T> arg, List<T> items, BindingRecyclerViewAdapterFactory factory, BindingRecyclerViewAdapter.ItemIds<T> itemIds) {
+        BindingRecyclerViewAdapter<T> adapter = (BindingRecyclerViewAdapter<T>) recyclerView.getAdapter();
+        if (adapter == null) {
+            adapter = factory.create(recyclerView, arg);
+            adapter.setItemIds(itemIds);
+            recyclerView.setAdapter(adapter);
+        }
+        adapter.setItems(items);
+    }
+
+    @BindingAdapter("layoutManager")
+    public static void setLayoutManager(RecyclerView recyclerView, LayoutManagers.LayoutManagerFactory layoutManagerFactory) {
+        recyclerView.setLayoutManager(layoutManagerFactory.create(recyclerView));
+    }
+
+    @BindingConversion
+    public static BindingRecyclerViewAdapterFactory toRecyclerViewAdapterFactory(final String className) {
+        return new BindingRecyclerViewAdapterFactory() {
+            @Override
+            public <T> BindingRecyclerViewAdapter<T> create(RecyclerView recyclerView, ItemViewArg<T> arg) {
+                return Utils.createClass(className, arg);
+            }
+        };
+    }
+}